feat(duckies): add deployment steps
