+++ contracts/package.json
@@ -8,10 +8,10 @@
     },
     "devDependencies": {
         "@nomicfoundation/hardhat-chai-matchers": "^1.0.1",
-        "@nomiclabs/hardhat-ethers": "^2.0.6",
-        "@nomiclabs/hardhat-etherscan": "^3.0.3",
-        "@openzeppelin/contracts-upgradeable": "^4.6.0",
-        "@openzeppelin/hardhat-upgrades": "^1.17.0",
+        "@nomiclabs/hardhat-ethers": "^2.1.0",
+        "@nomiclabs/hardhat-etherscan": "^3.1.0",
+        "@openzeppelin/contracts-upgradeable": "^4.7.0",
+        "@openzeppelin/hardhat-upgrades": "^1.19.0",
         "@typechain/ethers-v5": "^7.2.0",
         "@typechain/hardhat": "^2.3.1",
         "@types/chai": "^4.3.1",
@@ -30,8 +30,8 @@
         "eslint-plugin-prettier": "^3.4.1",
         "eslint-plugin-promise": "^5.2.0",
         "eth-gas-reporter": "^0.2.25",
-        "ethers": "^5.6.8",
-        "hardhat": "^2.9.7",
+        "ethers": "^5.6.9",
+        "hardhat": "^2.10.1",
         "hardhat-gas-reporter": "^1.0.8",
         "prettier": "^2.6.2",
         "prettier-plugin-solidity": "^1.0.0-beta.13",
