--- contracts/scripts/deploy.ts
@@ -1,28 +1,0 @@
-// We require the Hardhat Runtime Environment explicitly here. This is optional
-// but useful for running the script in a standalone fashion through `node <script>`.
-//
-// When running the script with `npx hardhat run <script>` you'll find the Hardhat
-// Runtime Environment's members available in the global scope.
-
-import { ethers, upgrades } from "hardhat";
-
-async function main() {
-  const [deployer] = await ethers.getSigners();
-
-  console.log("Deploying contracts with the account:", deployer.address);
-  console.log("Account balance:", (await deployer.getBalance()).toString());
-
-  const Duckies = await ethers.getContractFactory("Duckies");
-
-  const token = await upgrades.deployProxy(Duckies as any, [
-    process.env.SIGNER_ACCOUNT as string,
-  ]);
-
-  await token.deployed();
-  console.log("Token deployed to:", token.address);
-}
-
-main().catch((error) => {
-  console.error(error);
-  process.exitCode = 1;
-});
